update: 記録の編集画面を全てページに変更
- 編集画面はモーダルウィンドウで表示

- 新たにEditLog.js, AddLog.js, EditTimeLog.jsを作成してページを分割
