+++ src/App.css
@@ -9,96 +9,9 @@ h1, h2, h3, p, div {
   border-color: rgb(55,53,47);
 }
 
-.editRecordModalOverlay {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  right: 0;
-  left: 0;
-  width: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 50;
-}
-
-.editRecordModalContent {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: #fff;
-  width: 95%;
-  padding: 30px 15px;
-  outline: none;
-  box-sizing: border-box;
-  height: 400px;
-}
-
-.addRecordModalOverlay {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  right: 0;
-  left: 0;
-  width: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 50;
-}
-
-.addRecordModalContent {
-  width: 100%;
-  height: 500px;
-  background-color: #fff;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  padding: 30px 15px;
-  box-sizing: border-box;
-  outline: none;
-  overflow: scroll;
-}
-
-.modalContent {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 300px;
-  /* height: 150px; */
-  padding: 20px;
-  z-index: 70;
-  background-color: #fff;
-  outline: none;
-  border-radius: 10px;
-}
-
-.modalOverlay {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  right: 0;
-  left: 0;
-  width: 100%;
-  z-index: 60;
-  background-color: rgba(0, 0, 0, 0.5);
-}
-
 .border {
   border-right: 1px solid rgba(0, 0, 0, 0.12);
 }
-
-@media(min-width: 600px) {
-  .editRecordModalContent{
-    width: 370px;
-  }
-
-  .addRecordModalContent {
-    width: 580px;
-    height: 650px;
-    padding: 30px;
-  }
-}
-
 
 .time-box {
   position: absolute;
